Fix 'walks like a window, quacks like a window' test
In some browsers (at least Opera and IE) the inclusion of the prototype
library can cause the tested window properties to become available in
other HTML elements.

An effect of this is that the change introduced in [6313] causes the
width to be reported incorrectly on these elements (the document width
is used instead of the actual element width).

This is solved by checking for parentNode, which is null or undefined
on window and document objects but not on HTML elements.

Fixes #6088