Prevent ajaxLocParts exception on about:blank page.
When working with the Windows WebBrowser control, it is common to load
a HTML page from memory into an 'about:blank' page. Currently jQuery
throws the following exception when sourced from an 'about:blank' page:

   uncaught TypeError: Cannot read property '1' of null

The exception is from trying to access 'ajaxLocParts[ 1 ]', but ajaxLocParts
will be null on an 'about:blank' page because it does not match the 'rurl' regexp.