Slots for animations
When multiple animations are meant to be executed in "parallel",
currently they will be asynchronous and thus experience an delayed
animation for later elements. To overcome this issue, we add an slot
mechanism which specifies that we can use one single slot for a bunch of
animations, and they all will use the same start time and current time.

closes #6281