Fixes #9381.
When using a browser that supports "requestAnimationFrame", and firing off an animation, we call setTimeout to fire after the animation is finished, triggering fx.tick(). Thus, if requestAnimationFrame is never fired (when the window is hidden or minimized) we still see the animation triggered, with a maximum of one callback/DOM update per animation.

One can simulate the new behavior by commenting-out all calls to "requestAnimationFrame" in jQuery.