Fix timers deletion race condition.
On each fx tick, each timer step function in jQuery.timers is called to
check animation completion, and if so, the timer is deleted from the
array. If the animation complete callback calls a method that deletes
the timer itself (like stop), tick will then delete the wrong timer
pointed now by the loop counter, causing an unrelated element animation
to be stopped. To avoid it, check that the timer we are deleting after
the step function call is still the same object.
A test checking the correct animation result is added as well.
A fiddle demonstrating the problem can be found at
http://jsfiddle.net/aitorciki/LE9eT/ .