A fix for Ticket #9217 - jQuery crashing in IE.
This ticket describes a jquery crash in IE when an element being
animated is removed. Unfortunately, this fix uses a try/catch -
probably due to my lack of knowledge of the internals of IE.

From what I can infer, when an element is switched out from under
javascript, it is an error to write to scrollTop; the only way to
know whether it is an error to write to a field is to write to it:
thus the try/catch.