Let the props in $(html, props) be any jQuery.fn method.
This avoids special cases since we can eliminate attrFn. It does create some ambiguity since someone may add a plugin on jQuery.fn and change the meaning of one of the props. This does not bother me since I think this signature is an incredibly bad idea anyway. Rather than deprecate it, we can enhance it beyond repair!