Add back a sacrificial $.attrFn object.
Originally removed in 86b775d036627ebd7242fbb4eb9f24e4ba1fa9c5 as part of the enhancement that allows $(html, props) to use any $.fn method.

Although $.attrFn is undocumented it appears to be a poorly kept secret. jQuery Mobile 1.1 is using it and it's the topic of several blog/StackOverflow posts. Leave an empty object here as a dumpster for now, but it's coming out for good in 1.9.