Works around a crashing browser bug affecting Android browser < Android 2.3.5. The bug was triggered by code in effects. Fixes #12497.
A simple example (without jQuery) of JS that will trigger this crashing browser bug is here: http://jsfiddle.net/wCQvh/

This browser bug appears to have been resolved in Android 2.3.5. Unfortunately, there are a large number of active devices that cannot and/or will not be upgraded.

The workaround in this commit preserves the same logic as the previous code but avoids triggering the browser bug.