JSHint: Make compatible with jshint, lint test/data.
The "options" and "globals" properties are grunt-jshint specific.
Which means running node-jshint from cli, or using an editor with
jshint/jshintrc support would not read any of the options and thus
blow the editor full of errors (based on the default options in
jshint core).

Alphabetised the options again and separated the individual
groups with a line break (restriction, laxing, environment).

Especially because restriction is true-to-restrict, and laxing is
true-to-lax. Mixed was somewhat confusing.

Removed laxing options and globals that were no longer used.

Added a .jshintignore file to make sure that it actually passes
properly.