No ticket. Remove the unnecessary guard in addGetHookIf. Close gh-1426.
In 1.x if the support test executes before doc ready, it may not be able
to return a result yet. In such a case, we protect ourselves from future
breakages, allowing only for the ones before doc ready.

Since in 2.x lazy support tests attach test elements to docElem, not body,
such a guard is unnecessary.