Manipulation: Use textarea for missing IE defaultValue check
IE11 fixed the checkbox defaultValue issue but not textarea. Rather than
creating a new detect name I'm reusing the old one to protect anyone who
is unwisely using this externally. Re-fixing the defaultValue when it
doesn't need to be done is not a problem, so leave that code for IE11.

Fixes #14716
Closes gh-1495