Update val.js
Seeing a behavior change in jQuery('option').val() as reported in http://bugs.jquery.com/ticket/14858.  Behavior of .val() was changed which may return insignificant white space when a value attribute is not used.

Seems to have been changed because of this ticket:
http://bugs.jquery.com/ticket/14686
in this commit:
https://github.com/jquery/jquery/commit/92cbf5362cfe4fe52e919eecfebe8ecbe27ee5e3#diff-ac2456f6444b226fd09f497031b2f7d2

Since the problem is in IE10/11 and only for an empty value, can we simply return element.text.  If an exception is caught, return an empty string?