Update attributes.js
From https://github.com/jquery/jquery/pull/1531, currently discussing if the result from val() should be equivalent to textContent in this scenario. This test is to be consistent with 1.10.1 and 2.x branches.  It may be that	1.11.0 is the correct branch.

Given that the browser would show an empty option value for this (no white-space long), and that the w3.org spec on option has the same definition for an empty label and value, and that the length of label for the following snippet would be zero, I'll argue that the value shoudd also be a length of zero.