Fix to update if key was added as-is when using dash
When adding data using object argument, the keys are not camel cased as internally .extend is used. This fails when trying to update the value. http://jsfiddle.net/c0k0owkq/

The fix is to either use a different method instead of `.extend` to adding data from an object which uses camel cased keys or check if the key exist as-is when updating. The later was an easy change.