Manipulation: buildFragment always uses document of the context
buildFragment methods such as createDocumentFragment or createElement.
It called it on context, but the context could have been some random
node, not necesarilly document node. This commit fixes that method so it
always takes the ownerDocument of the context (or the context if it's
document itself) and callse the mentioned methods on that