toggle: renamed argument 'state' to 'doShow'
To make it clear that passing true shows and passing false hides. Otherwise if you take the documentation with showOrHide at its word, it would do nothing if you pass false and would call the regular toggle() if you pass true. 
I created a pull request for the api.jquery.change as well.
https://github.com/jquery/api.jquery.com/pull/654