Check before setting style property
According to [this jsperf](http://jsperf.com/check-style-attributes-before-setting-or-just-constant-), checking if a style property has changed before setting it gives a speedup of 40% (Chrome) to 80% (Firefox) if the value is the same. This check is also hard to to in userland because the user would have to replicate camel case, vendor prefixes, hooks etc. which are applied inside `jQuery.style`.