Allow passing ".class-name" to .addClass, .removeClass, .toggleClass, .hasClass
Many times it is useful to declare a class once and reuse it throughout a piece of code (e.g. plugins).

    var CUSTOM_CLASS = '.class-name',
        $el = $(CUSTOM_CLASS);

    ...

    if ($someEl.hasClass(CUSTOM_CLASS)) {
        $someEl.removeClass(CUSTOM_CLASS);
    }

    ...

    $someOtherEl.find(CUSTOM_CLASS);

    // etc

However, this is not possible currently since `.addClass`, `.removeClass`, `.toggleClass`, and `.hasClass` do not accept class names which include the dot, while `$()` and most tree traversing methods either accept or require it.

I think it would be beneficial to allow these other methods to accept class names which include the dot.