Alias ajax as http
You've already aliased `type` to `method`, and fixed other semantic "bugs". Why not go all the way? Kick the decade old misnomer in the face and embrace a strong and clear message to the rising generation of developers.

It's just so difficult to explain how x stands for "JSON" and "Binary Data", y'know?