Hidden attribute solution for show/hide/toggle
A simpler solution for show/hide/toggle based on the HTML5 hidden attribute.

To avoid ambiguity regarding how to write CSS to ensure hidden elements should be hidden, the hidden attribute value is set to hidden instead of empty.

This solution is more in line with best practices regarding how to show/hide elements, as it leaves it up to the author to ensure it's hidden properly via CSS (essentially the same as the class-based recommendation, removing/adding classes to control display). 

It will also leave the inline style intact, in case this is needed to overrule jQuery's toggling.

For responsive web sites, this will also provide a better way to control whether or not the element should be shown, as it can be controlled through the hidden attribute selector within the context being styled.

Not to mention: this probably performs better than having to remember the computed styles and resetting these when showing the element.

Another option is to add a method "unhide" that removes the hidden attribute, whereas "show" retains the functionality of the current code where its inline style is changed to force it to show.