Update deferred.js
So that `new jQuery.Deferred()` now returns a true instance of the class. (Calls to jQuery.Deferred() as a factory directly remains untouched)

Update ajax.js

Let jqXHR pickup the changes made to jQuery.Deferred so that it is now an instance of it instead of a plain object returned by .promise();

Added unit test case.