Deferred: Fix exceptions silent swallowing
This fix applies to the usage of ready() calls,
in order to allow exception catching when someone
throws an uncatched exception inside the ready function body.
This is solved by giving the ability to pass a handling function
argument in the ready call, that will be called once the exception
is thrown.

The fix works on the ready implementation,
not the ready-not-deferred one.

Fixes #3174