Deferred: Fix exceptions silent swallowing (Update)
In order for errors of "Error" type to be printed, the
regex in the deferred.exceptionHook() needed a slight modification,
adding the '?' character for the Error prefixes.

Additionally the deferred.exceptionHook() function will print
the process stack trace only if not undefined, something that
could be issued as another bug, but it is now fixed.

The fix works on the ready implementation,
not the ready-not-deferred one.

Fixes #3174