Make more robust for odd environments
Not sure if the `global.window.document` bit is needed (seems like checking `global.document` would make more sense), but suppose it can't hurt. The comments (here and on previous pull requests) seem to imply that there may not be a `global` document in CommonJS (would certainly be odd to have `window` without `document`). For the same reason, am now passing `global.window` instead of global. Think this would all be more clear if window and document were passed separately as the `window.document` references make little sense in a browser.

Haven't checked to see if jQuery uses this reference where the global object would make more sense. If so, then it would make sense to pass the global object as well.