Put back window reference and 'use strict'
We are painted into a corner here as the factory only receives one `window` object, which is definitely *not* the same as the global object in NodeJS. As a result, it will add `$` and `jQuery` to something other than the global object in NodeJS (e.g. the jsdom `window` object).

Best I can do at the moment. You really need to pass two references to your factory: `window` and global and use each appropriately (e.g. this code should likely use the global object, not `window`).

HTH