Deferred: Use typeof instead of isFunction
jQuery's Deferred is expected to be a close approximation to Promise of
ECMAScript 2016 Language Specification. Promise defined in ECMAScript 2016
tests onFulfilled, onRejected and then are objects with [[Call]].

However, isFunction, used in the Deferred, tests toStringTag, which is
irrelevant from the definition. As the result, Deferred shows different
behaviors for some objects such as GeneratorFunction.

This fix uses typeof operator, which will test the given value is an
objects with [[Call]], instead.