Tests: Ensure all urls to data use baseURI
Various urls were missing this prefix, either explicitly or via url().
This commit fixes the odd ones that were missing the prefix, which
was causing some tests to fail under Karma runner, where the base
is different because tests are served from root (/), but files
are served from /base (e.g. /base/test/data).

Also, to avoid regressing this in the future, change baseURI to
be the path to /test/data instead of the path to /test so that
it doesn't silently make tests without it pass when run from
/test/index.html.