Improved trim performance in Selector.js
The native JavaScript trim() method works much faster than regular expressions in the case of removing whitespace from the beginning and the end of a string.

The trim using the regular expression of selector.js is the same as Sizzle, but it seems desirable to remove it for the following reasons.
-The current version and the later of IE9 supports String.prototype.trim. jQuery V2.0 has stopped supporting the prior versions in IE8.
-Since Sizzle is removed from 4.0, there is no need to keep the code.

However, since regular expressions for trim are also used for other purposes, the only portion of the codes with the improved performance with the trim () method are changed.