Docs: Update Frequently Reported Issues in the GitHub issue template
The issue about selectors with '#' being broken is old and no longer frequently
reported so this commit removes it from the list. On the other hand, we're now
getting lots of reports about the security fix in jQuery 3.5.0 that was also
a breaking change: gh-4642. This one is now mentioned in the list.

Ref gh-4642